pyproject.toml
[project]
name = "langgraph_recipes"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.11, <4.0"
dependencies = [
    "pydantic (>=2.10.6,<3.0.0)",
    "langchain-openai (>=0.3.2,<0.4.0)",
    "langchain-anthropic (>=0.3.4,<0.4.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "langgraph (>=0.2.68,<0.3.0)",
    "langchain-community (>=0.3.16,<0.4.0)"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
packages = [
    { include = "utils" }
]
